+++ lib/class/podcast.class.php
@@ -323,7 +323,7 @@ class Podcast extends database_object implements library_item
      */
     public static function create(array $data, $return_id = false)
     {
-        $feed = $data['feed'];
+        $feed = (string) $data['feed'];
         // Feed must be http/https
         if (strpos($feed, "http://") !== 0 && strpos($feed, "https://") !== 0) {
             AmpError::add('feed', T_('Feed URL is invalid'));
@@ -352,6 +352,15 @@ class Podcast extends database_object implements library_item
         $lastbuilddate = 0;
         $episodes      = array();
         $arturl        = '';
+
+        // don't allow duplicate podcasts
+        $sql = "SELECT `id` FROM `podcast` WHERE `feed`= ?";
+        $db_results = Dba::read($sql, $feed);
+        while ($row = Dba::fetch_assoc($db_results, false)) {
+            if ((int) $row['id'] > 0) {
+                return (int) $row['id'];
+            }
+        }
 
         $xmlstr = file_get_contents($feed, false, stream_context_create(Core::requests_options()));
         if ($xmlstr === false) {